Update tool pair Tags case insensitive
There was a problem with Tags - "Postgresql" in Tackle 1.2, "PostgreSQL"
in Tackle 2.0. The pairing code which should migrate just Tags which are
not already present in Tackle 2.0 consideted these to be a different
tags. However sqlite prevented creation of the "Postgresql" tag when there was
"PostgreSQL" alredy present in the same TagType.

Adding lower() to Tag names to tackle tool to match tags case
insensitive.

Export/import of whole Tackle 1.2 with -s option was not affected by this
problem since all seed tags were deleted from Tackle2 before import.

https://issues.redhat.com/browse/TACKLE-683
Signed-off-by: Marek Aufart <maufart@redhat.com>