Add GitHub CODEOWNERS file
Adding CODEOWNERS file to notify QE when swagger API spec files get changed
(several updates in past caused QE broken tooling, this is a simple way
to get them notified that there is something new in API spec).

If we want, there could be core team to be notified about all changes,
but based on our current process, we look on Hub PRs anyway, so it is
not needed much probably.

Signed-off-by: Marek Aufart <maufart@redhat.com>